refactor: change models and services interfaces to client
