+++ MyTheFourth/src/MyTheFourth.Frontend/MyTheFourth.Frontend.Web.Client/Constants/MyTheFourthHttpServicePath.cs
@@ -1,4 +1,4 @@
-namespace MyTheFourth.Frontend.Web.Components;
+namespace MyTheFourth.Frontend.Web.Client.Constants;
 
 public static class MyTheFourthHttpServicePath
 {
