+++ MyTheFourth/src/MyTheFourth.Frontend/MyTheFourth.Frontend.Web.Client/Models/Character.cs
