+++ MyTheFourth/src/MyTheFourth.Frontend/MyTheFourth.Frontend.Web.Client/_Imports.razor
@@ -8,3 +8,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using MyTheFourth.Frontend.Web.Client
+@using MyTheFourth.Frontend.Web.Client.Services.Interfaces
+@using MyTheFourth.Frontend.Web.Models
