+++ MyTheFourth/src/MyTheFourth.Frontend/MyTheFourth.Frontend.Web/DependencyInjections/HttpClientDependencyInjectionExtensions.cs
@@ -1,42 +1,48 @@
-
+using MyTheFourth.Frontend.Web.Client.Services;
 using MyTheFourth.Frontend.Web.Http.Handlers;
-using MyTheFourth.Frontend.Web.Services;
+
+namespace MyTheFourth.Frontend.Web.DependencyInjections;
 
 public static class HttpClientDependencyInjectionExtensions
 {
-    public static IServiceCollection AddApiHttpService(this IServiceCollection services){
+    public static IServiceCollection AddApiHttpService(this IServiceCollection services)
+    {
 
-        services.AddHttpClient<MyTheFourthHttpService>("Default", (httpclient) => {
+        services.AddHttpClient<MyTheFourthHttpService>("Default", (httpclient) =>
+        {
             httpclient.BaseAddress = new Uri("http://");
         }).AddHttpMessageHandler(provider => provider.GetService<ApiDelegatingHandler>()!);
 
-            var implementationType = typeof(MyTheFourthHttpService);
+        var implementationType = typeof(MyTheFourthHttpService);
 
-            var interfaces = implementationType.GetInterfaces();
+        var interfaces = implementationType.GetInterfaces();
 
-            foreach (var type in interfaces)
-            {
-                services.AddTransient(type, provider => provider.GetRequiredService<MyTheFourthHttpService>());
-            }
+        foreach (var type in interfaces)
+        {
+            services.AddTransient(type, provider => provider.GetRequiredService<MyTheFourthHttpService>());
+        }
 
-            services.AddTransient<ApiDelegatingHandler>();
+        services.AddTransient<ApiDelegatingHandler>();
 
-            return services;
+        return services;
     }
 
-     public static IServiceCollection AddApiHttpServiceFake(this IServiceCollection services){
+    public static IServiceCollection AddApiHttpServiceFake(this IServiceCollection services)
+    {
+        services.AddScoped<MyTheFourthHttpServiceFake>();
 
-            var implementationType = typeof(MyTheFourthHttpService);
-
-            var interfaces = implementationType.GetInterfaces();
-
-            foreach (var type in interfaces)
-            {
-                services.AddTransient(type, provider => provider.GetRequiredService<MyTheFourthHttpService>());
-            }
+        var implementationType = typeof(MyTheFourthHttpServiceFake);
 
 
-            return services;
+        var interfaces = implementationType.GetInterfaces();
+
+        foreach (var type in interfaces)
+        {
+            services.AddTransient(type, provider => provider.GetRequiredService<MyTheFourthHttpServiceFake>());
+        }
+
+
+        return services;
     }
