+++ MyTheFourth/src/MyTheFourth.Frontend/MyTheFourth.Frontend.Web/Services/MyTheFourthHttpServiceFake.cs
@@ -1,5 +1,5 @@
+using MyTheFourth.Frontend.Web.Client.Services.Interfaces;
 using MyTheFourth.Frontend.Web.Models;
-using MyTheFourth.Frontend.Web.Services.Interfaces;
 
 public class MyTheFourthHttpServiceFake :
 IMoviesService,
